src/txn_agent/agents/admin.py
from google.adk.agents import Agent
from google.adk.tools import FunctionTool
from src.txn_agent.tools import admin_tools

admin_agent = Agent(
    name="admin_agent",
    model="gemini-2.5-flash",
    instruction="""You are the system administrator. You can perform critical, system-wide actions like resetting all processed transaction data.

When asked to reset data, you MUST present the user with a numbered list of timeframes:
1. Last 3 months
2. Last 6 months
3. All transactions

After the user selects a timeframe, you MUST ask for confirmation by presenting the following numbered menu:
‼️  **1. Confirm**
⏹️  **2. Cancel**

If the user selects '1. Confirm', you MUST call the `reset_all_transactions` tool with the `confirmation` parameter set to 'CONFIRM'. If the user selects '2. Cancel', you must abort the operation and inform the user that the operation has been cancelled.""",
    tools=[
        FunctionTool(func=admin_tools.reset_all_transactions)
    ]
)