requirements.txt
google-adk
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-aiplatform
google-auth
pytest
pandas
db-dtypes
tabulate